+++ internal/item/service.go
@@ -2,7 +2,13 @@ package item
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+
+	"github.com/goverland-labs/platform-events/events/core"
+	"github.com/rs/zerolog/log"
+
+	"github.com/goverland-labs/feed/internal/subscriber"
 )
 
 //go:generate mockgen -destination=mocks_test.go -package=item . DataProvider,Publisher
@@ -15,25 +21,91 @@ type DataProvider interface {
 	Create(item FeedItem) error
 }
 
-type Service struct {
-	repo   DataProvider
-	events Publisher
+type SubscriberProvider interface {
+	GetByID(_ context.Context, id string) (*subscriber.Subscriber, error)
 }
 
-func NewService(r DataProvider, p Publisher) (*Service, error) {
+type SubscriptionProvider interface {
+	GetSubscribers(_ context.Context, daoID string) ([]string, error)
+}
+
+type Service struct {
+	repo          DataProvider
+	events        Publisher
+	subscribers   SubscriberProvider
+	subscriptions SubscriptionProvider
+}
+
+func NewService(r DataProvider, p Publisher, sub SubscriberProvider, sp SubscriptionProvider) (*Service, error) {
 	return &Service{
-		repo:   r,
-		events: p,
+		repo:          r,
+		events:        p,
+		subscribers:   sub,
+		subscriptions: sp,
 	}, nil
 }
 
-func (s *Service) HandleItem(_ context.Context, item FeedItem) error {
+func (s *Service) HandleItem(ctx context.Context, item FeedItem) error {
 	err := s.repo.Create(item)
 	if err != nil {
 		return fmt.Errorf("can't create item: %w", err)
 	}
 
-	// todo: publish core event
+	// todo: refactor and move to separated logic
+	go func() {
+		subs, err := s.subscriptions.GetSubscribers(ctx, item.DaoID)
+		if err != nil {
+			log.Error().Err(err).Msg("get subscribers")
+			return
+		}
+
+		feed := convertToExternalFeed(item)
+		data, err := json.Marshal(feed)
+		if err != nil {
+			log.Error().Err(err).Msgf("marshal feed: %d", item.ID)
+			return
+		}
+
+		for _, sub := range subs {
+			info, err := s.subscribers.GetByID(ctx, sub)
+			if err != nil {
+				log.Error().Err(err).Msgf("get details for: %s", sub)
+				continue
+			}
+
+			payload := core.CallbackPayload{
+				WebhookURL: info.WebhookURL,
+				Body:       data,
+			}
+
+			err = s.events.PublishJSON(ctx, core.SubjectCallback, payload)
+			if err != nil {
+				log.Error().Err(err).Msgf("publish callback for: %s", info.WebhookURL)
+			}
+		}
+	}()
 
 	return nil
+}
+
+func convertFeedType(ftype Type) core.Type {
+	switch ftype {
+	case TypeDao:
+		return core.TypeDao
+	case TypeProposal:
+		return core.TypeProposal
+	default:
+		return core.TypeDao
+	}
+}
+
+func convertToExternalFeed(item FeedItem) core.FeedItem {
+	return core.FeedItem{
+		DaoID:        item.DaoID,
+		ProposalID:   item.ProposalID,
+		DiscussionID: item.DiscussionID,
+		Type:         convertFeedType(item.Type),
+		Action:       core.ConvertActionToExternal(item.Action),
+		Snapshot:     item.Snapshot,
+	}
 }