+++ internal/subscription/models.go
@@ -1,0 +1,9 @@
+package subscription
+
+import "gorm.io/gorm"
+
+type Subscription struct {
+	gorm.Model
+	SubscriberID string
+	DaoID        string
+}