+++ pkg/grpcsrv/server.go
@@ -1,13 +1,14 @@
 package grpcsrv
 
 import (
+	grpcauth "github.com/grpc-ecosystem/go-grpc-middleware/auth"
 	"google.golang.org/grpc"
 )
 
-func NewGrpcServer() *grpc.Server {
+func NewGrpcServer(excludePath []string, auth grpcauth.AuthFunc) *grpc.Server {
 	server := grpc.NewServer(
-		StdUnaryMiddleware(),
-		StdStreamMiddleware(),
+		StdUnaryMiddleware(UnaryReflectionFilter(excludePath, grpcauth.UnaryServerInterceptor(auth))),
+		StdStreamMiddleware(StreamReflectionFilter(excludePath, grpcauth.StreamServerInterceptor(auth))),
 	)
 
 	StdRegister(server)